# docs/swagger.yaml
basePath: /
definitions:
...
  model.Certs:
    properties:
      certificate:
        type: string
      expires_at:
        type: string
      key:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: LoRaWAN Platform Connector API
paths:
...
  /gateways/{hub_id}/cert:
    post:
      description: Generates a new certificate
      parameters:
      - description: Hub ID
        in: path
        name: hub_id
        required: true
        type: string
      responses:
        "200":
          description: generated certificate
          schema:
            $ref: '#/definitions/model.Certs'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Generate Certificate
      tags:
      - Gateways
...
swagger: "2.0"